Don't bring the interface back up ourselves
Even though it takes a while, we have to let coreos re-up the interface
to properly enslave it.

This is a *workaround* style fix, there are major bugs in the way
coreos-cloudinit is handling the Debian-style interfaces file, causing
eth1 to never get configured properly. We should address the root cause
of the issue later.